Fix movement amount type.

+++ src/main/java/UTN/FRGP/TP_L5_GRUPO_1/Models/Movement.java
@@ -72,11 +72,11 @@ public class Movement {
         this.movementType = movementType;
     }
 
-    public Integer getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
